Move files to commons directory
