+++ commons/ui/src/main/kotlin/com/teacherworkout/commons/ui/composables/SearchView.kt
@@ -1,4 +1,4 @@
-package com.teacherworkout.features.learn
+package com.teacherworkout.commons.ui.composables
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.shape.RoundedCornerShape
@@ -11,12 +11,12 @@ import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Close
 import androidx.compose.material.icons.filled.Search
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.MutableState
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.TextFieldValue
+import com.teacherworkout.commons.ui.R
 
 @Composable
 fun SearchView(
