+++ commons/ui/src/main/kotlin/com/teacherworkout/commons/ui/data/LessonsRepository.kt
@@ -1,0 +1,7 @@
+package com.teacherworkout.commons.ui.data
+
+import com.teacherworkout.commons.ui.model.LessonTheme
+
+interface LessonsRepository {
+    suspend fun getAllLessonThemes(): Result<List<LessonTheme>>
+}
