+++ commons/ui/src/main/kotlin/com/teacherworkout/commons/ui/data/Result.kt
@@ -1,6 +1,6 @@
-package com.teacherworkout.features.learn.data
+package com.teacherworkout.commons.ui.data
 
-sealed class Result<out R> {
+sealed class Result<out T> {
     data class Success<out T>(val data: T) : Result<T>()
     data class Error(val exception: Exception) : Result<Nothing>()
 }
