+++ commons/ui/src/main/res/values-ro-rRO/strings.xml
@@ -5,4 +5,5 @@
     <string name="title_learn">Învață</string>
     <string name="title_profile">Profil</string>
     <string name="title_discover">Descoperă</string>
+    <string name="title_themes">Categorii de lecții</string>
 </resources>
