+++ commons/ui/src/main/res/values/strings.xml
@@ -5,4 +5,5 @@
     <string name="title_learn">Learn</string>
     <string name="title_discover">Discover</string>
     <string name="title_profile">Profile</string>
+    <string name="title_themes">Lesson themes</string>
 </resources>
