--- features/learn/src/main/kotlin/com/teacherworkout/features/learn/data/LessonsRepository.kt
@@ -1,7 +1,0 @@
-package com.teacherworkout.features.learn.data
-
-import com.teacherworkout.features.learn.model.LessonTheme
-
-interface LessonsRepository {
-    suspend fun getAllLessonThemes(): Result<List<LessonTheme>>
-}
