+++ features/learn/src/main/kotlin/com/teacherworkout/features/learn/di/KoinModules.kt
@@ -1,6 +1,6 @@
 package com.teacherworkout.features.learn.di
 
-import com.teacherworkout.features.learn.data.impl.FakeLessonsRepository
+import com.teacherworkout.commons.ui.data.impl.FakeLessonsRepository
 import com.teacherworkout.features.learn.discover.DiscoverViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
