+++ features/learn/src/main/res/values-ro-rRO/strings.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <string name="hello_blank_fragment">Hello themes fragment</string>
-    <string name="themes_title">Categorii de lecții</string>
     <string name="discover_lessons_title">Descoperă lecțiile</string>
     <string name="search_label">Caută un subiect</string>
 </resources>
